backend/requirements.txt
asgiref==3.6.0
astroid==2.15.3
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
dill==0.3.6
Django==4.1.6
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
docopt==0.6.2
Faker==18.3.1
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.3
mccabe==0.7.0
mrjob==0.7.4
pipreqs==0.4.11
platformdirs==3.2.0
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.7
tzdata==2022.7
urllib3==1.26.14
wrapt==1.15.0
yarg==0.1.9
